move setJsonData out of delayed to scripts
