+++ scripts/delayed.js
@@ -126,24 +126,3 @@ if (jsonLdMeta) {
   addLdJsonScript(document.querySelector('head'), jsonLdMeta.content);
   document.querySelector('meta[name="json-ld"]').remove();
 }
-
-/**
- * Set the JSON-LD script in the head
- * @param {*} data
- * @param {string} name
- */
-// eslint-disable-next-line import/prefer-default-export
-export function setJsonLd(data, name) {
-  const existingScript = document.head.querySelector(`script[data-name="${name}"]`);
-  if (existingScript) {
-    existingScript.innerHTML = JSON.stringify(data);
-    return;
-  }
-
-  const script = document.createElement('script');
-  script.type = 'application/ld+json';
-
-  script.innerHTML = JSON.stringify(data);
-  script.dataset.name = name;
-  document.body.appendChild(script);
-}
